% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{apply_double_gates}
\alias{apply_double_gates}
\title{Apply double marker gates}
\usage{
apply_double_gates(data1, double_gates_df)
}
\arguments{
\item{data1}{Data frame with single marker gates}

\item{double_gates_df}{Data frame with marker1, marker2 columns for double gates}
}
\value{
Data frame with double gates added
}
\description{
Apply double marker gates
}
\keyword{internal}
