% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{apply_single_gates}
\alias{apply_single_gates}
\title{Apply single marker gates}
\usage{
apply_single_gates(data1, gate_thresholds, slide_name)
}
\arguments{
\item{data1}{Data frame with marker measurements}

\item{gate_thresholds}{Data frame with gating thresholds}

\item{slide_name}{Name of the current slide}
}
\value{
Data frame with gating results added as logical columns
}
\description{
Apply single marker gates
}
\keyword{internal}
