% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{clean_formula_label}
\alias{clean_formula_label}
\title{Create clean column names from formula expressions}
\usage{
clean_formula_label(expr)
}
\arguments{
\item{expr}{A formula expression}
}
\value{
Character string with cleaned name
}
\description{
Create clean column names from formula expressions
}
\keyword{internal}
