% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{create_splitting_line}
\alias{create_splitting_line}
\title{Create a splitting line through overlapping region using Voronoi diagram}
\usage{
create_splitting_line(intersection_sf, two_polygon_sf)
}
\arguments{
\item{intersection_sf}{The intersection geometry between two polygons}

\item{two_polygon_sf}{The two original polygons}
}
\value{
An sf linestring geometry for splitting, or NULL if failed
}
\description{
Create a splitting line through overlapping region using Voronoi diagram
}
\keyword{internal}
