% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{create_summary}
\alias{create_summary}
\title{Create summary statistics by grouping variables}
\usage{
create_summary(data, group_vars)
}
\arguments{
\item{data}{Data frame with cell measurements}

\item{group_vars}{Character vector of column names to group by}
}
\value{
Data frame with summary statistics
}
\description{
Create summary statistics by grouping variables
}
\keyword{internal}
