% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{cycif_save_results}
\alias{cycif_save_results}
\title{Save results to files}
\usage{
cycif_save_results(results, output_dir)
}
\arguments{
\item{results}{Result list from \code{cycif_summarize()}}

\item{output_dir}{Directory to save files (will be created if it doesn't exist)}
}
\value{
NULL (called for side effects)
}
\description{
Saves the results from \code{cycif_summarize()} to CSV and QS files.
}
\examples{
\dontrun{
results <- cycif_summarize(sampled_data)
cycif_save_results(results, "/path/to/output")
}
}
