% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{expand_roi_boundaries}
\alias{expand_roi_boundaries}
\title{Expand ROI boundaries to include nearby cells}
\usage{
expand_roi_boundaries(data1, expand_distance)
}
\arguments{
\item{data1}{Data frame with ROI assignments}

\item{expand_distance}{Distance in microns to expand ROIs}
}
\value{
Data frame with expanded ROI assignments
}
\description{
Expand ROI boundaries to include nearby cells
}
\keyword{internal}
