% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{extend_linestring}
\alias{extend_linestring}
\title{Extend a linestring by a given distance at both ends}
\usage{
extend_linestring(linestring, distance)
}
\arguments{
\item{linestring}{A linestring geometry}

\item{distance}{Distance to extend at each end}
}
\value{
Extended linestring geometry
}
\description{
Extend a linestring by a given distance at both ends
}
\keyword{internal}
