% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{generate_overall_recommendation}
\alias{generate_overall_recommendation}
\title{Generate overall recommendation based on overlap analysis}
\usage{
generate_overall_recommendation(
  small_overlaps,
  large_overlaps,
  multiway_overlaps,
  containment_cases
)
}
\arguments{
\item{small_overlaps}{Number of small overlaps}

\item{large_overlaps}{Number of large overlaps}

\item{multiway_overlaps}{Number of multi-way overlaps}

\item{containment_cases}{Number of containment cases}
}
\value{
Character string with overall recommendation
}
\description{
Generate overall recommendation based on overlap analysis
}
\keyword{internal}
