% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{parse_point_str}
\alias{parse_point_str}
\title{Parse ROI points from string format}
\usage{
parse_point_str(roi_points_str)
}
\arguments{
\item{roi_points_str}{String with ROI points in the format "x1,y1 x2,y2 ..."}
}
\value{
Matrix with parsed points (x, y)
}
\description{
Parse ROI points from string format
}
\keyword{internal}
