% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{resolve_pairwise_overlap}
\alias{resolve_pairwise_overlap}
\title{Resolve overlap between exactly two polygons}
\usage{
resolve_pairwise_overlap(two_polygon_sf)
}
\arguments{
\item{two_polygon_sf}{An sf object with exactly 2 polygon geometries}
}
\value{
An sf object with the two polygons after resolving their overlap
}
\description{
Resolve overlap between exactly two polygons
}
\keyword{internal}
