% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{standardize_coordinates}
\alias{standardize_coordinates}
\title{Standardize coordinate columns and convert to microns}
\usage{
standardize_coordinates(data1, scale_factor)
}
\arguments{
\item{data1}{Data frame with cell measurements}

\item{scale_factor}{Scale factor for converting pixels to microns}
}
\value{
Data frame with standardized coordinate columns
}
\description{
Standardize coordinate columns and convert to microns
}
\keyword{internal}
