% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{standardize_input_data}
\alias{standardize_input_data}
\title{Standardize input data formats}
\usage{
standardize_input_data(data)
}
\arguments{
\item{data}{Input data: list of dataframes, single dataframe with slideName,
or single dataframe without slideName}
}
\value{
List of dataframes by slide name
}
\description{
Converts input data to a standardized list of dataframes by slide
}
\examples{
\dontrun{
# Single dataframe with slideName column
std_data <- standardize_input_data(my_data)

# List of dataframes
std_data <- standardize_input_data(list(slide1 = df1, slide2 = df2))
}
}
