% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_roi_metadata}
\alias{add_roi_metadata}
\title{Add ROI metadata to summary}
\usage{
add_roi_metadata(summary_data, roi_metadata)
}
\arguments{
\item{summary_data}{Data frame with summary statistics}

\item{roi_metadata}{Data frame with ROI metadata}
}
\value{
Data frame with metadata joined
}
\description{
Add ROI metadata to summary
}
