% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_slide_metadata}
\alias{add_slide_metadata}
\title{Add slide metadata to summary}
\usage{
add_slide_metadata(data, slide_metadata)
}
\arguments{
\item{data}{Data frame with with \code{SlideName} column}

\item{slide_metadata}{Data frame with slide metadata}
}
\value{
Data frame with metadata joined
}
\description{
Add slide metadata to summary
}
