% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{find_polygon_overlaps}
\alias{find_polygon_overlaps}
\title{Find all overlapping polygon pairs}
\usage{
find_polygon_overlaps(polygon_sf)
}
\arguments{
\item{polygon_sf}{An sf object with polygon geometries}
}
\value{
An sf dataframe with intersection geometries and area columns
}
\description{
Find all overlapping polygon pairs
}
\keyword{internal}
